Add cmd/gopherjs-ng binary that wraps around the `go` tool.
In theory, this approach should allow to keep only compiler and linker
parts of GopherJS and rely on the mainline Go to handle all the detail
of the build system, such as modules, file embedding, etc.

This tool work by setting -toolexec flag to point at "gopherjs-ng
adaptor", which then acts as if it was the standard compiler/linker. For
the moment it doesn't actually do anything, but `go build -n` works
successfully.

As of now, this needs a custom version of the go tool, which disables
GOOS/GOARCH pair validity check, since GopherJS uses GOARCH=js. Ideally,
we could get a way to disable that check.