Implement basic logic for constructing a virtual GOROOT.
This is an alternative way of applying GopherJS augmentations,
materializing them on disk. Standard library packages that are require
no augmentations are symlinked into the destination GOROOT. Packages
that do have augmentations are passed through a set of filters:

 - First, only relevant original files are kept. For example, all
   assembly is excluded.
 - Then, the remaining original files are processed and all symbols that
   exist are a subject to augmentation are removed from the original
   sources
 - Finally, augmentation sources are also added to the package to make
   it complete again.

I added a few simplified integration tests for this logic, which are
designed to be agnostic of optimizations such as symlinks and focus on
the correctness of the content instead.